Time: 3 ms (28.81%) | Memory: 8.6 MB (53.89%) - LeetSync

+++ 7-reverse-integer/reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int x) {
-        long long rev=0;
+         long rev=0;
         while(x){
             rev=rev*10+x%10;
             x/=10;
